extract ProjectSettingsProvider
